Admin_Dashboard/server/app/auth/jwt.py
"""JWT and password utilities – use same secret_key as main server if sharing tokens."""
from datetime import datetime, timedelta
from typing import Any
import bcrypt
from jose import jwt
from app.config import settings


def hash_password(password: str) -> str:
    secret = password.encode("utf-8")
    if len(secret) > 72:
        secret = secret[:72]
    return bcrypt.hashpw(secret, bcrypt.gensalt()).decode("utf-8")


def verify_password(plain: str, hashed: str) -> bool:
    try:
        return bcrypt.checkpw(
            plain.encode("utf-8")[:72],
            hashed.encode("utf-8"),
        )
    except Exception:
        return False


def create_access_token(data: dict[str, Any]) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=settings.access_token_expire_minutes)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)


def decode_access_token(token: str) -> dict | None:
    try:
        return jwt.decode(token, settings.secret_key, algorithms=[settings.algorithm])
    except Exception:
        return None
